Create exported package to validate metadata in TypeScript #1: move json files to schemes folder
