+++ js-traits-validator/index.js
@@ -6,8 +6,8 @@ Object.defineProperty(exports, "__esModule", { value: true });
 exports.DEFAULT_REGISTRY = exports.METADATA_SCHEMA = exports.validateTraits = exports.validateMetadata = void 0;
 const _2020_1 = __importDefault(require("ajv/dist/2020"));
 const ajv_formats_1 = __importDefault(require("ajv-formats"));
-const registry_json_1 = __importDefault(require("./scripts/registry.json"));
-const metadata_schema_json_1 = __importDefault(require("../traits/metadata_schema.json"));
+const registry_json_1 = __importDefault(require("./schemes/registry.json"));
+const metadata_schema_json_1 = __importDefault(require("./schemes/metadata_schema.json"));
 const ajv = new _2020_1.default();
 (0, ajv_formats_1.default)(ajv);
 const metadataValidate = ajv.compile(metadata_schema_json_1.default);
