+++ js-traits-validator/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@traittech/traits-validator",
-  "version": "0.0.1",
+  "version": "0.0.2",
   "main": "index.js",
   "types": "index.d.ts",
   "license": "Apache-2.0",
