+++ js-traits-validator/schemes/registry.json
